==> 04-ebullition-analysis.Rmd <==
# Ebullition Analysis

I want to evaluate the relationship between ebullitive flux of CH4 and depth and distance to the inlet. I hypothesized that both depth and distance to the inlet will be negatively associated with ebullitive flux. Ultimately, we have to deal with the fact that the data are zero-inflated and bounded by 0. I tried multiple models below to address these characteristics
```{r eb analysis set up, warning=FALSE}
eb_all <- bind_rows(sister3_pts_22_eb %>% mutate(Site ="Sister3"), 
                    pick_pts_13_eb,
                    catfish_pts_18_eb %>% dplyr::select(-c(catfish_dist_1, catfish_dist_2)),
                    catfish_pts_06_eb %>% dplyr::select(-c(catfish_dist_1, catfish_dist_2)),
                    deans_pts_30_eb %>% dplyr::select(-c(deans_dist_1, deans_dist_2)),
                    deans_pts_16_eb%>% dplyr::select(-c(deans_dist_1, deans_dist_2))) %>% 
  mutate(Site = as.factor(Site))%>% 
  mutate(Depth_scale = scale(Depth), inlet_dist_scale = scale(inlet_dist))

ebullition_24 <- eb_all %>% 
  mutate(ch4_eb_g = gas_vol_corrected * (1/10^6) * ch4_conc * (1/10^3) * 16.04 * (1/0.0821)* (1/293.15), 
         Collection.Start.Date = case_when(
           Sampling == "Deans1" ~ ymd("2022-08-16"),
           Sampling == "Deans2" ~ ymd("2022-08-30"), 
           Sampling == "Catfish1" ~ ymd("2022-09-06"),
           Sampling == "Catfish2" ~ ymd("2022-09-18"), 
           Sampling == "Pick1" ~ ymd("2022-09-13"),
           Sampling == "Sister31" ~ ymd("2022-08-22")
         )) %>% 
  mutate(time_hr_na = if_else(is.na(ch4_eb_g), NA_real_, time_hr)) %>% 
  group_by(Site, Trap, Sampling, Collection.Start.Date, inlet_dist, Depth, inlet_dist_scale, Depth_scale) %>% 
  summarize(sum_time = sum(time_hr_na, na.rm = T), sum_g = sum(ch4_eb_g, na.rm = T)) %>% 
  mutate(area = pi * 81 * 0.0001, ch4_eb_g_m2_hr_24 = sum_g * (1/area)* (1/sum_time))
```

## Plots
```{r dist and depth plot, echo = F}
# Depth and inlet distance are correlated, VIF of mod1 = 1.2
cor.test(eb_all$Depth, eb_all$inlet_dist, method = "spearman")

ggplot(eb_all, aes(Depth, ch4_eb_g_m2_hr, color = Site)) + geom_point()
ggplot(eb_all, aes(inlet_dist, ch4_eb_g_m2_hr, color = Site)) + geom_point()

```

## Mod1: Non-zero data, fixed effects only
```{r fixed non zero, echo =F}
# no random effects and no zeros because gamma can't handle zeros
mod1 <- glm(ch4_eb_g_m2_hr ~ Depth_scale + inlet_dist_scale, data = eb_all %>% filter(ch4_eb_g_m2_hr > 0), family = Gamma(link = "log"))
summary(mod1)
```

## Mod2: Non zero data, mixed model
```{r mixed non zero, echo = F}
# with random effects but still no zeros
mod2 <- glmer(ch4_eb_g_m2_hr ~ Depth_scale + inlet_dist_scale + (1|Sampling)  + (1|trip_wr), data = eb_all %>% filter(ch4_eb_g_m2_hr > 0), family = Gamma(link = "log"))
summary(mod2)

sd(eb_all %>% filter(ch4_eb_g_m2_hr>0) %>% pull(Depth), na.rm=T)
sd(eb_all %>% filter(ch4_eb_g_m2_hr>0) %>% pull(inlet_dist), na.rm=T)
```

## Mod2.a: Non zero data, mixed model without Picks
```{r no pick mixed non zero, echo = F}
mod2.a <- glmer(ch4_eb_g_m2_hr ~ Depth_scale + inlet_dist_scale + (1|Sampling)  + (1|trip_wr), data = eb_all %>% filter(ch4_eb_g_m2_hr > 0, Site != "Picks"), family = Gamma(link = "log"))
summary(mod2.a)

sd(eb_all %>% filter(ch4_eb_g_m2_hr>0, Site != "Picks") %>% pull(Depth), na.rm=T)
sd(eb_all %>% filter(ch4_eb_g_m2_hr>0, Site != "Picks") %>% pull(inlet_dist), na.rm=T)
```

## Mod2.b: Non zero data, mixed model only Picks
```{r pick mixed non zero, echo = F}
mod2.b <- glmer(ch4_eb_g_m2_hr ~ Depth_scale + inlet_dist_scale + (1|trip_wr), data = eb_all %>% filter(ch4_eb_g_m2_hr > 0, Site == "Picks"), family = Gamma(link = "log"))
summary(mod2.b)
```

## Mod3: Logistic regression, fixed effects only
```{r logistic fixed, echo = F}
mod3 <- glm(ch4_eb_bin ~ Depth_scale + inlet_dist_scale, data = eb_all %>% mutate(ch4_eb_bin = if_else(ch4_eb_g_m2_hr>0, 1,0)), family = binomial(link = "logit"))
summary(mod3)
```

## Mod4: Logistic regression, mixed model
```{r logistic mixed, echo = F}
mod4 <- glmer(ch4_eb_bin ~ Depth_scale + inlet_dist_scale + (1|Sampling) + (1|trip_wr), data = eb_all %>% mutate(ch4_eb_bin = if_else(ch4_eb_g_m2_hr>0, 1,0)), family = binomial(link = "logit"))
summary(mod4)
```

## Mod4.a: Logistic regression, mixed model without Picks
```{r no pick logistic mixed, echo = F}
mod4.a <- glmer(ch4_eb_bin ~ Depth_scale + inlet_dist_scale + (1|Sampling) + (1|trip_wr), data = eb_all %>% filter(Site != "Picks") %>% mutate(ch4_eb_bin = if_else(ch4_eb_g_m2_hr>0, 1,0)), family = binomial(link = "logit"))
summary(mod4.a)
```

## Mod4.b: Logistic regression, mixed model only Picks
```{r pick logistic, echo = F}
mod4.b <- glm(ch4_eb_bin ~ Depth_scale + inlet_dist_scale, data = eb_all %>% filter(Site == "Picks") %>% mutate(ch4_eb_bin = if_else(ch4_eb_g_m2_hr>0, 1,0)), family = binomial(link = "logit"))
summary(mod4.b)
```

## Mod5: Hurdle model, predictors generate structural zeros
```{r hurdle pred, echo = F}
mod5 <- glmmTMB(ch4_eb_g_m2_hr ~ Depth_scale + inlet_dist_scale + (1|Sampling)  + (1|trip_wr), data = eb_all %>% filter(is.na(ch4_eb_g_m2_hr)==F) %>% st_drop_geometry(),  ziformula = ~., family = ziGamma)
summary(mod5)
```

## Model selection with Mod5 predictors
```{r model selection hurdle, echo = F}
mod5.a <-  glmmTMB(ch4_eb_g_m2_hr ~ Depth_scale + (1|Sampling)  + (1|trip_wr), data = eb_all %>% filter(is.na(ch4_eb_g_m2_hr)==F) %>% st_drop_geometry(),  ziformula = ~., family = ziGamma)
mod5.b <- glmmTMB(ch4_eb_g_m2_hr ~ inlet_dist_scale + (1|Sampling)  + (1|trip_wr), data = eb_all %>% filter(is.na(ch4_eb_g_m2_hr)==F) %>% st_drop_geometry(),  ziformula = ~., family = ziGamma)

cand.models <- list("mod5"=mod5, 
                    "mod5.a"=mod5.a,
                    "mod5.b"=mod5.b)

aictab(cand.models)

summary(mod5.b)
```

## Mod6: Hurdle model, equal probability of structural zeros
```{r hurdle no pred, echo = F}
mod6 <- glmmTMB(ch4_eb_g_m2_hr ~ Depth_scale + inlet_dist_scale + (1|Sampling)  + (1|trip_wr), data = eb_all %>% filter(is.na(ch4_eb_g_m2_hr)==F) %>% st_drop_geometry(),  ziformula = ~1, family = ziGamma)
summary(mod6)
```

## Mod7: Not broken up by time, no longer zero-inflated, sampling random effect
```{r aggregated gamma, echo = F}
mod7 <- glmer(ch4_eb_g_m2_hr_24 ~ Depth + inlet_dist + (1|Sampling) , data = ebullition_24 %>% filter(ch4_eb_g_m2_hr_24 > 0), family = Gamma(link = "log"))
summary(mod7)
r2_nakagawa(mod7)

mod7.a <- glmer(ch4_eb_g_m2_hr_24 ~ inlet_dist + (1|Sampling) , data = ebullition_24 %>% filter(ch4_eb_g_m2_hr_24 > 0), family = Gamma(link = "log"))
summary(mod7.a)
r2_nakagawa(mod7.a)

mod7.b <- glmer(ch4_eb_g_m2_hr_24 ~ Depth + (1|Sampling) , data = ebullition_24 %>% filter(ch4_eb_g_m2_hr_24 > 0), family = Gamma(link = "log"))
summary(mod7.b)
r2_nakagawa(mod7.b)


cor(ebullition_24 %>% select(Depth, inlet_dist, ch4_eb_g_m2_hr_24) %>% st_drop_geometry(), use = "complete.obs")
cor(ebullition_24 %>% select(Depth, inlet_dist, ch4_eb_g_m2_hr_24) %>% st_drop_geometry(), use = "complete.obs", method = "spearman")

cand.models <- list("mod7"= mod7, 
                    "mod7.a"= mod7.a,
                    "mod7.b"= mod7.b)
```

## By transect
```{r}
mod8 <- glmer(ch4_eb_g_m2_hr_24 ~ Transect + (1|Sampling), data = ebullition_24 %>% mutate(Transect = str_sub(Trap, 2, 2) %>% as.numeric()) %>% filter(ch4_eb_g_m2_hr_24 > 0), family = Gamma(link = "log"))
summary(mod8)
r2_nakagawa(mod8)
```


```{r plot depth gamma, echo = F}
sampling.labs <- c("Catfish\n9/6 - 9/7", "Catfish\n9/18 - 9/19", "Deans\n8/16 - 8/17", "Deans\n8/30 - 8/31", "Blue Herron\n9/13 - 9/14", "Sister\n8/22 - 8/23")
names(sampling.labs) <- c("Catfish1", "Catfish2", "Deans1", "Deans2", "Pick1", "Sister31")

pred.mod7.b <- ggpredict(mod7.b, terms =c("Depth", "Sampling"), type = "random") 
pred.mod7.b

names(ebullition_24)

depth_fig <- ggplot(pred.mod7.b) +
  geom_ribbon(aes(x = x, ymin = conf.low, ymax = conf.high), alpha = 0.2)+
  geom_point(data=ebullition_24 %>% rename("group" = "Sampling"), 
             aes(x=Depth, y = ch4_eb_g_m2_hr_24, group = group),
             size = 2.5, shape = 1, stroke = 1)+
   geom_line(aes(x=x, y=predicted), lwd = 2, color = "#94A7C5")+
  ylab(expression(CH[4]~ flux ~ (g~m^-2~hr^-1)))+
  xlab("Depth (m)")+
  theme_bw()+
  coord_cartesian(ylim = c(0,0.1))+
  theme(legend.title = element_blank(), axis.text = element_text(size = 14, color = "black"), axis.title = element_text(size = 16, color = "black"), legend.text = element_text(size = 14), strip.text = element_text(size = 14,  color = "black"))+
  facet_grid(~group, labeller = labeller(group = sampling.labs))

pred.mod7.a <- ggpredict(mod7.a, terms =c("inlet_dist", "Sampling"), type = "random") 

dist_fig <- ggplot(pred.mod7.a) +
 geom_ribbon(aes(x = x, ymin = conf.low, ymax = conf.high), alpha = 0.2)+
  geom_point(data=ebullition_24 %>% rename("group" = "Sampling"), 
             aes(x=inlet_dist, y = ch4_eb_g_m2_hr_24, group = group),
             size = 2.5, shape = 1, stroke = 1)+
   geom_line(aes(x=x, y=predicted), lwd = 2, color = "#94A7C5")+
  ylab(expression(CH[4]~ flux ~ (g~m^-2~hr^-1)))+
  xlab("Distance from the inlet (m)")+
  theme_bw()+
  coord_cartesian(ylim = c(0,0.1))+
  theme(legend.title = element_blank(), axis.text = element_text(size = 14, color = "black"), axis.title = element_text(size = 16, color = "black"), legend.text = element_text(size = 14), strip.text = element_text(size = 14,  color = "black"))+
  facet_grid(~group, labeller = labeller(group = sampling.labs))

#png(filename = "4-Figures/eb-analysis-summary.png", width = 8, height = 8, units = "in", res = 300)
plot_grid(depth_fig, dist_fig, labels = "AUTO", nrow = 2)
#dev.off()
```

```{r eb analysis rm, include = F}
to_remove <- c("deans_pts_16_eb", "deans_pts_30_eb", "pick_pts_13_eb", "sister3_pts_22_eb", "catfish_pts_06_eb", "catfish_pts_18_eb", "cand.models", "mod1", "mod2", "mod2.a", "mod2.b", "mod3", "mod4", "mod4.a", "mod4.b", "mod5", "mod5.a", "mod5.b", "mod6", "mod7", "mod7.a", "mod7.b", "mod8", "pred.mod7.a", "pred.mod7.b")
rm(list = to_remove)
```
