==> 08-total-estimates.Rmd <==
# Total flux estimates

```{r total estimates set up, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

all_fluxes_pathways <- date_from_sampling(read.csv("5-Results/1-kriged-estimates/all_fluxes_pathways.csv") %>% select(-X)%>% mutate(site = str_sub(sampling, 1, -2)))

all_fluxes_periods <- read.csv("5-Results/1-kriged-estimates/all_fluxes_periods.csv")%>% 
  select(-X) %>% 
  mutate(flux_mean_g_m2_hr = mean_flux_g_hr/area, 
         flux_2.5_g_m2_hr = flux_2.5_g_hr/area, 
         flux_97.5_g_m2_hr = flux_97.5_g_hr/area)
```

## Summary flux table masses and rates
```{r summary flux pathways, echo = F}
all_fluxes_pathways  %>% select(site, date, pathway, gas, flux_mean_g_m2_day:flux_97.5_g_day_co2_eq)  

all_fluxes_pathways  %>% select(site, date, pathway, gas, flux_mean_g_m2_day) %>% mutate(flux_round = round(flux_mean_g_m2_day, digits = 4)) %>% select(-flux_mean_g_m2_day) %>%  pivot_wider(names_from = c(pathway, gas), values_from = "flux_round")  %>% select(site, date, diffusion_CO2, diffusion_CH4, ebullition_CH4) %>% mutate(total_ch4 = diffusion_CH4 + ebullition_CH4)

all_fluxes_pathways  %>% select(site, date, pathway, gas, flux_2.5_g_m2_day) %>% mutate(flux_round = round(flux_2.5_g_m2_day, digits = 4)) %>% select(-flux_2.5_g_m2_day) %>%  pivot_wider(names_from = c(pathway, gas), values_from = "flux_round") %>% select(site, date, diffusion_CO2, diffusion_CH4, ebullition_CH4)

all_fluxes_pathways  %>% select(site, date, pathway, gas, flux_97.5_g_m2_day) %>% mutate(flux_round = round(flux_97.5_g_m2_day, digits = 4)) %>% select(-flux_97.5_g_m2_day) %>%  pivot_wider(names_from = c(pathway, gas), values_from = "flux_round") %>% select(site, date, diffusion_CO2, diffusion_CH4, ebullition_CH4)
```

## Summary Table CO2-eq
```{r co2 eq table, echo = F}
#flux_mean_g_m2_day_co2_eq
all_fluxes_pathways %>% 
  mutate(flux_mean_g_m2_day_co2_eq =  flux_mean_g_day_co2_eq/area,
         flux_2.5_g_m2_day_co2_eq= flux_2.5_g_day_co2_eq/area, 
         flux_97.5_g_m2_day_co2_eq= flux_97.5_g_day_co2_eq/area)%>% 
  select(site, date, pathway, gas, flux_mean_g_m2_day_co2_eq) %>% 
  pivot_wider(id_cols = c("site", "date"), names_from = c("pathway", "gas"), values_from = "flux_mean_g_m2_day_co2_eq") %>% 
  mutate(total = diffusion_CH4 + diffusion_CO2 + ebullition_CH4, prop_ch4_diff = (diffusion_CH4/total)*100, prop_ch4_eb = (ebullition_CH4/total)*100, prop_co2 = (diffusion_CO2/total)*100, prop_ch4 = prop_ch4_diff + prop_ch4_eb)
```

```{r summary co2 eq table, include = F}
green_comp <- all_fluxes_pathways %>% 
  group_by(sampling, site, date) %>% 
  mutate(flux_mean_g_m2_day_co2_eq =  flux_mean_g_day_co2_eq/area,
         flux_2.5_g_m2_day_co2_eq= flux_2.5_g_day_co2_eq/area, 
         flux_97.5_g_m2_day_co2_eq= flux_97.5_g_day_co2_eq/area) %>% 
  summarize(co2_eq_total_mean = sum(flux_mean_g_m2_day_co2_eq), co2_eq_total_2.5 = sum(flux_2.5_g_m2_day_co2_eq), co2_eq_total_97.5 = sum(flux_97.5_g_m2_day_co2_eq)) %>% 
  select(site, date, co2_eq_total_mean:co2_eq_total_97.5) %>% 
  mutate(green = if_else(site == "Pick", "green", "not green")) %>% 
  group_by(green) %>% 
  summarize(mean_co2_eq = mean(co2_eq_total_mean))
  
clarke_ponds <- st_read("2-Clean Data/clarke_ponds.shp") %>% filter(AreSqKm < 0.1)
clarke_area <- sum(clarke_ponds$AreSqKm)

green_avg <- green_comp$mean_co2_eq[1]
not_green_avg <- green_comp$mean_co2_eq[2]

df <- data.frame(intercept = (sum(clarke_ponds$AreSqKm)* not_green_avg), slope = ((sum(clarke_ponds$AreSqKm)* green_avg) - (sum(clarke_ponds$AreSqKm)* not_green_avg)))

ggplot(df, aes(intercept, slope)) +theme_bw()+xlab("Proportion of green ponds") + ylab("MG CO2-eq/day")+geom_abline(aes(intercept = intercept, slope = slope)) + ylim(c(df$intercept, (df$intercept+df$slope)))+xlim(c(0,1))

green_avg/not_green_avg

temp <- all_fluxes_pathways %>% 
  group_by(sampling, site, date) %>% 
  mutate(flux_mean_g_m2_day_co2_eq =  flux_mean_g_day_co2_eq/area,
         flux_2.5_g_m2_day_co2_eq= flux_2.5_g_day_co2_eq/area, 
         flux_97.5_g_m2_day_co2_eq= flux_97.5_g_day_co2_eq/area) %>% 
  summarize(co2_eq_total_mean = sum(flux_mean_g_m2_day_co2_eq), co2_eq_total_2.5 = sum(flux_2.5_g_m2_day_co2_eq), co2_eq_total_97.5 = sum(flux_97.5_g_m2_day_co2_eq)) %>% 
  select(site, date, co2_eq_total_mean:co2_eq_total_97.5) %>% 
  mutate(green = if_else(site == "Pick", "green", "not green"))

green_avg / temp$co2_eq_total_mean
```

### Pathway

```{r pathways summary fig, echo = F}
#png(filename = "4-Figures/co2-eq-summary.png", width = 8, height = 5, units = "in", res = 300)
all_fluxes_pathways %>% 
  mutate(flux_mean_g_m2_day_co2_eq =  flux_mean_g_day_co2_eq/area,
         flux_2.5_g_m2_day_co2_eq= flux_2.5_g_day_co2_eq/area, 
         flux_97.5_g_m2_day_co2_eq= flux_97.5_g_day_co2_eq/area, 
         pathway_gas = paste(pathway, gas, sep = " ") %>% as.factor() %>% fct_relevel("diffusion CO2", after = 0), 
         site_date = paste(site, date, sep = " ")) %>% 
  ggplot(aes(site_date, flux_mean_g_m2_day_co2_eq, fill = pathway_gas)) + 
  geom_bar(position = position_dodge(), stat = "identity")+
  scale_x_discrete(labels = c("Catfish 09/06/22" = "Catfish 06 Sept", "Catfish 09/18/22" = "Catfish 18 Sept", "Deans 08/16/22" = "Deans 16 Aug", "Deans 08/30/22" = "Deans 30 Aug", "Pick 09/13/22" = "Blue Herron 13 Sept", "Sister3 08/22/22" = "Sister 22 Aug"))+
  geom_errorbar(aes(ymax=flux_97.5_g_m2_day_co2_eq, ymin=flux_2.5_g_m2_day_co2_eq), position = position_dodge())+
  scale_fill_manual(values = c("#BFDBC0","#94A7C5", "#3D5170"), labels = c(expression(diffusion~CO[2]), expression(diffusion~CH[4]), expression(ebullition~CH[4])))+
  ylab(expression(CO[2]-equivalent ~ flux ~ (g~m^-2~day^-1)))+
  theme_bw()+
  theme(legend.text.align = 0, axis.title.x = element_blank(), legend.title = element_blank(), axis.text.x = element_text(angle=45, hjust=1), axis.text = element_text(size = 14, color = "black"), axis.title = element_text(size = 16, color = "black"), legend.text = element_text(size = 14))
#dev.off()

#png(filename = "4-Figures/co2-eq-summary-gas-comparison.png", width = 8.5, height = 5, units = "in", res = 300)
all_fluxes_pathways %>% 
  mutate(flux_mean_g_m2_day_co2_eq =  flux_mean_g_day_co2_eq/area,
         flux_2.5_g_m2_day_co2_eq= flux_2.5_g_day_co2_eq/area, 
         flux_97.5_g_m2_day_co2_eq= flux_97.5_g_day_co2_eq/area, 
         pathway_gas = paste(pathway, gas, sep = " ") %>% as.factor() %>% fct_relevel("diffusion CO2", after = 0),
         site = if_else(site == "Pick", "Blue Herron", site),
         site_date = paste(site, date, sep = " ")) %>% 
  ggplot(aes(site_date, flux_mean_g_m2_day_co2_eq, fill=gas)) + 
  geom_bar(position = position_dodge(), stat = "identity")+
  scale_fill_manual(values = c("#BFDBC0","#94A7C5"))+
  ylab(expression(CO[2]-equivalent ~ flux ~ (g~m^-2~day^-1)))+
  theme_bw()+
  theme(axis.title.x = element_blank(), legend.title = element_blank(), axis.text.x = element_text(angle=45, hjust=1), axis.text = element_text(size = 14, color = "black"), axis.title = element_text(size = 16, color = "black"), legend.text = element_text(size = 14))
#
dev.off()

```

## Diel Diffusive CH4 
Untransformed rate
```{r diel ch4 rate, echo = F}
#png(filename = "4-Figures/ch4-diffusion-periods.png", width = 9, height = 5, units = "in", res = 300)
ch4_diel <- add_sampling_period(all_fluxes_periods %>% 
  filter(!(sampling == "Deans1" & period == 8)) %>% 
  filter(gas == "CH4", pathway == "diffusion")) %>% 
  ggplot(aes(period, flux_mean_g_m2_hr))+
  geom_point()+
  geom_errorbar(aes(ymin = flux_2.5_g_m2_hr, ymax = flux_97.5_g_m2_hr))+
  geom_line()+
  facet_wrap(~sampling)+
  scale_x_discrete(limits = factor(c(1,2,3,4,5,6,7,8)), labels = c("10:00", "13:00", "16:00", "19:00", "22:00", "01:00", "04:00", "07:00"))+
  theme_bw()+
  ylab(expression(Diffusive ~ Flux ~ (g~CH[4]~m^-2~hr^-1)))+
  theme(axis.title.x = element_blank(), 
        axis.text = element_text(size=14, color = "black"), 
        axis.text.x = element_text(angle = 45, hjust = 1), 
        axis.title.y = element_text(size=16, color = "black"), 
        strip.text = element_text(size=14, color = "black"))
ch4_diel
#dev.off()
```

Untransformed without Picks
```{r diel ch4 no pick, echo = F}
add_sampling_period(all_fluxes_periods %>% 
  filter(gas == "CH4", pathway == "diffusion", site != "Pick")) %>% 
  ggplot(aes(period, flux_g_mean))+
  geom_point()+
  geom_errorbar(aes(ymin = flux_g_2.5, ymax = flux_g_97.5))+
  geom_line()+
  facet_wrap(~sampling)+
  scale_x_discrete(limits = factor(c(1,2,3,4,5,6,7,8)), labels = c("10AM", "1PM", "4PM", "7PM", "10PM", "1AM", "4AM", "7AM"))+
  theme_bw()+
  ylab("Diffusive Flux (g CH4/day)")+
  theme(axis.title.x = element_blank())
```

## Diel Diffusive CO2 

```{r diel co2 rate, echo = F}
#Deans 16 final sampling was actually at 9 AM so excluding

#png(filename = "4-Figures/co2-diffusion-periods.png", width = 9, height = 5, units = "in", res = 300)
co2_diel <- add_sampling_period(all_fluxes_periods %>% 
  filter(gas == "CO2", pathway == "diffusion") %>% 
  filter(!(sampling == "Deans1" & period == 8))) %>% 
  ggplot(aes(period, flux_mean_g_m2_hr))+
  geom_point()+
  geom_errorbar(aes(ymin = flux_2.5_g_m2_hr, ymax = flux_97.5_g_m2_hr))+
  geom_line()+
  facet_wrap(~sampling)+
  scale_x_discrete(limits = factor(c(1,2,3,4,5,6,7,8)), labels = c("10:00", "13:00", "16:00", "19:00", "22:00", "01:00", "04:00", "07:00"))+
  theme_bw()+
  ylab(expression(Diffusive ~ Flux ~ (g~CO[2]~m^-2~hr^-1)))+
  theme(axis.title.x = element_blank(), 
        axis.text = element_text(size=14, color = "black"), 
        axis.text.x = element_text(angle = 45, hjust = 1), 
        axis.title.y = element_text(size=16, color = "black"), 
        strip.text = element_text(size=14, color = "black"))
#dev.off()

co2_diel
```

```{r diel co2 rate no pick, echo = F}
#png(filename = "4-Figures/co2-diffusion-periods-no-pick.png", width = 9, height = 5, units = "in", res = 300)
add_sampling_period(all_fluxes_periods %>% 
  filter(gas == "CO2", pathway == "diffusion", site != "Pick") %>% 
  filter(!(sampling == "Deans1" & period == 8))) %>% 
  ggplot(aes(period, flux_mean_g_m2_hr))+
  geom_point()+
  geom_errorbar(aes(ymin = flux_2.5_g_m2_hr, ymax = flux_97.5_g_m2_hr))+
  geom_line()+
  facet_wrap(~sampling)+
  scale_x_discrete(limits = factor(c(1,2,3,4,5,6,7,8)), labels = c("10:00", "13:00", "16:00", "19:00", "22:00", "01:00", "04:00", "07:00"))+
  theme_bw()+
  ylab(expression(Diffusive ~ Flux ~ (g~CO[2]~m^-2~hr^-1)))+
  theme(axis.title.x = element_blank(), 
        axis.text = element_text(size=14, color = "black"), 
        axis.text.x = element_text(angle = 60, hjust = 1), 
        axis.title.y = element_text(size=16, color = "black"), 
        strip.text = element_text(size=14, color = "black"))
#dev.off()
```

## Diel ebullition
Untransformed rate
```{r diel eb rate, echo = F}
#png(filename = "4-Figures/ch4-ebullition-periods.png", width = 9, height = 5, units = "in", res = 300)
eb_diel <- all_fluxes_periods %>% 
    filter(gas == "CH4", pathway == "ebullition", period != "8~9") %>% 
   mutate(trip = case_when(period == "0~1" ~ 1,
                             period == "1~2" ~ 2, 
                             period == "2~3" ~ 3, 
                             period == "3~4" ~ 4, 
                             period == "4~5" ~ 5, 
                             period == "5~6" ~ 6, 
                             period == "6~7" ~ 7, 
                             period == "7~8" ~ 8)) %>% 
  mutate(sampling = case_when(sampling == "Catfish1" ~"Catfish 06-07 Sept",
                              sampling == "Catfish2" ~ "Catfish 18-19 Sept",
                              sampling == "Deans1" ~ "Deans 16-17 Aug",
                              sampling == "Deans2" ~ "Deans 30-31 Aug", 
                              sampling == "Pick1" ~ "Blue Herron 13-14 Sept",
                              sampling == "Sister31" ~ "Sister 22-23 Aug")) %>% 
  ggplot(aes(trip, flux_mean_g_m2_hr))+
  geom_point()+
  geom_errorbar(aes(ymin = flux_2.5_g_m2_hr, ymax = flux_97.5_g_m2_hr))+
  geom_line()+
  facet_wrap(~sampling)+
  scale_x_discrete(limits = factor(c(1,2,3,4,5,6,7,8)), labels = c("10:00", "13:00", "16:00", "19:00", "22:00", "01:00", "04:00", "07:00"))+
  theme_bw()+
  ylab(expression(Ebullitive ~ Flux ~ (g~CH[4]~m^-2~hr^-1)))+
  theme(axis.title.x = element_blank(), 
        axis.text = element_text(size=14, color = "black"), 
        axis.text.x = element_text(angle = 45, hjust = 1), 
        axis.title.y = element_text(size=16, color = "black"), 
        strip.text = element_text(size=14, color = "black"))
#dev.off()

eb_diel
```

```{r include = F}
# all_fluxes_periods %>% filter(gas == "CO2") %>% 
#   mutate(tod = case_when(period %in% c(1,2,3,8) ~ "day",
#                          period %in% c(4,5,6,7) ~ "night")) %>% 
#   group_by(sampling, tod) %>% 
#   summarize(mean_flux_sum = sum(flux_g_mean)) %>% 
#   pivot_wider(id_cols = sampling, names_from = tod, values_from = mean_flux_sum) %>% mutate(tod_diff = day/night)
```

```{r knit diel figs, echo = F}
png(filename = "4-Figures/periods.png", width = 7.5, height = 11.5, units = "in", res = 300)
plot_grid(co2_diel, ch4_diel, eb_diel, ncol = 1, labels = "AUTO", label_size = 18, vjust = 1.3)
dev.off()
```
