==> 12-diel-analyses.Rmd <==
# Analyses of patterns of diel diffusive emissions

```{r diel setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE)

source("data_munge_functions.R")

par <- read.csv("2-Clean Data/par-data.csv")
ws <- date_to_sampling(read.csv("2-Clean Data/windspeeds.csv") %>% mutate(date = ymd(date)))
lux <- sampling_filter_24hr(site_rename(read.csv("2-Clean Data/lux.csv") %>% mutate(Date_Time = ymd_hms(Date_Time, tz = "America/New_York")) %>% distinct()))

time_join <- join_by(site, closest(x$start_time >= y$Date_Time))

diffusion_env <- site_rename(diffusion) %>% 
  left_join(lux, time_join)%>% 
  select(site, location, start_time, trip, flux_ch4_g_hr, flux_co2_g_hr, Lux) 

time_join <- join_by(site, closest(x$start_time >= y$date))

diffusion_env <- diffusion_env %>% left_join(top_temp, time_join)

diffusion_env <- diffusion_env %>% 
  left_join(ws %>% mutate(sampling = case_when(
    sampling == "Catfish 9/6 - 9/7" ~ "Catfish 06-07 Sept",
    sampling == "Catfish 9/18 - 9/19" ~ "Catfish 18-19 Sept",
    sampling == "Deans 8/16 - 8/17" ~ "Deans 16-17 Aug",
    sampling == "Deans 8/30 - 8/31"  ~ "Deans 30-31 Aug",
    sampling == "Blue Herron 9/13 - 9/14"  ~ "Blue Herron 13-14 Sept",
    sampling == "Sister 8/22 - 8/23" ~ "Sister 22-23 Aug"
  )) %>% select(-date, -site), by = c("sampling", "location", "trip")) %>% 
  drop_na()

lux_daily <- sampling_filter_24hr(lux) %>% filter(is.na(sampling)==F)

time_vec <- c("10:00", "13:00", "16:00", "19:00", "22:00", "01:00", "04:00", "07:00")
date_time_vec_1 <- do.call(c, map(time_vec[1:5], function(x) ymd_hm(paste(date(Sys.time()), x, sep = " "), tz = "America/New_York")))
date_time_vec_2 <- do.call(c, map(time_vec[6:8], function(x) ymd_hm(paste(date(Sys.time()+days(1)), x, sep = " "), tz = "America/New_York")))
date_time_vec <- c(date_time_vec_1, date_time_vec_2)

png(filename = "4-Figures/lux-sampling-summary.png", width = 8, height = 5, units = "in", res = 300)
daily_trace_aes(lux_daily %>% 
     mutate(time = format(Date_Time, format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 0:9), time + days(1), time)) %>% 
  ggplot(aes(time_alt, Lux, group = sampling)) +
  geom_line()+
  ylab("Lux"), date_time_vec)
dev.off()
```

## Diel patterns in environmental variables
```{r env var plots}
time_vec <- c("10:00", "13:00", "16:00", "19:00", "22:00", "01:00", "04:00", "07:00")
date_time_vec_1 <- do.call(c, map(time_vec[1:5], function(x) ymd_hm(paste(date(Sys.time()), x, sep = " "), tz = "America/New_York")))
date_time_vec_2 <- do.call(c, map(time_vec[6:8], function(x) ymd_hm(paste(date(Sys.time()+days(1)), x, sep = " "), tz = "America/New_York")))
date_time_vec <- c(date_time_vec_1, date_time_vec_2)

# lux plot
daily_trace_aes(diffusion_env %>% 
  mutate(time = format(start_time, format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 1:9), time + days(1), time)) %>% 
  ggplot(aes(time_alt, Lux, group = sampling))+
  geom_point(), date_time_vec)

# ws plot
daily_trace_aes(diffusion_env %>% 
  filter(is.na(ws_mean)==F, is.na(sampling)==F) %>% 
  mutate(time = format(start_time, format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 1:9), time + days(1), time)) %>% 
  ggplot(aes(time_alt, ws_mean, group = sampling))+
  geom_point()+
  ylab("Windspeed (m/s)"), date_time_vec)

# temp plot
daily_trace_aes(diffusion_env %>% 
  filter(is.na(temp)==F, is.na(sampling)==F) %>% 
  mutate(time = format(start_time, format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 1:9), time + days(1), time)) %>% 
  ggplot(aes(time_alt, temp, group = sampling))+
  geom_point()+
  ylab("Temperature at 0.25m Depth (C)"), date_time_vec)
```

### Weather station hourly windspeed from Aug 14 - Sept 19
```{r weather station ws}
bot_ws <- read.csv("2-Clean Data/weatherstem_windspeed.csv") %>% mutate(date_time = mdy_hm(X, tz = "America/New_York"))

time_vec <- c("10:00", "13:00", "16:00", "19:00", "22:00", "01:00", "04:00", "07:00")
date_time_vec_1 <- do.call(c, map(time_vec[1:5], function(x) ymd_hm(paste(date(Sys.time()), x, sep = " "), tz = "America/New_York")))
date_time_vec_2 <- do.call(c, map(time_vec[6:8], function(x) ymd_hm(paste(date(Sys.time()+days(1)), x, sep = " "), tz = "America/New_York")))
date_time_vec <- c(date_time_vec_1, date_time_vec_2)

# this ws is in miles per hour
bot_ws %>% 
  mutate(time = format(date_time, format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 0:9), time + days(1), time)) %>% 
  ggplot(aes(time_alt, Anemometer, group = time_alt))+
  geom_boxplot()+
  #geom_point()+
  theme(axis.text = element_text(angle = 45, hjust = 1, size = 14, color = "black"))+
  scale_x_datetime(date_labels = "%H:%M", breaks = date_time_vec)

temporary <- sampling_filter_24hr_other(bot_ws %>% rename("Date_Time" = "date_time")) %>% filter(is.na(sampling)==F)

bot_ws.xts <- as.xts(bot_ws$Anemometer, order.by = bot_ws$date_time)
dygraph(bot_ws.xts)

1911.1- (118.11*25) + (3.4527*(25^2)) -(0.04132*(25^3))
1897.8- (114.28*25) + (3.2902*(25^2)) -(0.039061*(25^3))
```

### Dissolved oxygen 
```{r ysi do}
ysi <- read.csv("2-Clean Data/ysi_all.csv") %>% mutate(sampling = case_when(
  sampling == "Catfish1" ~ "Catfish 06-07 Sept",
  sampling == "Catfish2" ~ "Catfish 18-19 Sept",
  sampling == "Deans1" ~ "Deans 16-17 Aug",
  sampling == "Deans2" ~ "Deans 30-31 Aug",
  sampling == "Pick1" ~ "Blue Herron 13-14 Sept",
  sampling == "Sister31" ~ "Sister 22-23 Aug"
)) %>% rename("site" = "Location", "location"="Trap..", "trip"= "Trip") %>% 
  mutate(do = as.numeric(value), time = format(ymd_hms(date_time), format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 0:9), time + days(1), time)) %>% 
  mutate(do = replace(do, X == 2371, 4.59), do = replace(do, X==2707, 5.91), do = replace(do, X == 1832, NA), do = replace(do, X==852, 5.86)) 

daily_trace_aes(ysi %>% 
  filter(name == "YSI Bottom DO mg/L") %>% 
  mutate(do = as.numeric(value), time = format(ymd_hms(date_time), format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 0:9), time + days(1), time)) %>% 
  mutate(do = replace(do, X == 2371, 4.59), do = replace(do, X==2707, 5.91)) %>% # fix data entry errors
  ggplot(aes(time_alt, do, group = time_alt))+
  geom_point()+
  ylab("Bottom Dissolved Oxygen (mg/L)"), date_time_vec)


daily_trace_aes(ysi %>% 
  filter(name == "YSI Top DO mg/L") %>% 
  mutate(do = as.numeric(value), time = format(ymd_hms(date_time), format = "%H:%M") %>% as.POSIXct(format = "%H:%M"), time_alt = if_else(hour(time) %in% c(24, 0:9), time + days(1), time)) %>% 
  mutate(do = replace(do, X == 1832, NA), do = replace(do, X==852, 5.86))%>% # fix data entry errors
  ggplot(aes(time_alt, do, group = time_alt))+
  geom_point()+
  ylab("Top Dissolved Oxygen (mg/L)"), date_time_vec)

daily_trace_aes(ysi %>% 
  filter(name == "YSI Bottom DO mg/L" | name == "YSI Top DO mg/L") %>% 
  mutate(Trap = str_sub(location, 2, 3)) %>% 
  ggplot(aes(time_alt, do, group = time_alt))+
  geom_point(aes(color = name, shape = Trap), size = 2, stroke = 1)+
  ylab("Dissolved Oxygen (mg/L)")+
  scale_shape_manual(values = c(0,1,2,3)), date_time_vec)

diffusion_env <- diffusion_env %>%  left_join(ysi %>% filter(name == "YSI Bottom DO mg/L" | name == "YSI Top DO mg/L") %>% select(sampling, location, trip, name, do) %>%  pivot_wider(id_cols = c(sampling, location, trip), names_from = name, values_from = do) %>% group_by(sampling, trip) %>% summarize(mean_top_do = mean(`YSI Top DO mg/L`, na.rm = T), mean_bottom_do = mean(`YSI Bottom DO mg/L`, na.rm = T)) , by = c("sampling", "trip")) %>% mutate(start_date = case_when(
  sampling == "Catfish 06-07 Sept" ~ ymd("2022-09-06"),
  sampling == "Catfish 18-19 Sept" ~ ymd("2022-09-18"),
  sampling == "Deans 16-17 Aug" ~ ymd("2022-08-16"),
  sampling == "Deans 30-31 Aug" ~ ymd("2022-08-30"),
  sampling == "Blue Herron 13-14 Sept" ~ ymd("2022-09-13"),
  sampling == "Sister 22-23 Aug" ~ ymd("2022-08-23")
))

cor.test(diffusion_env$temp, diffusion_env$mean_top_do)
```

## CH4 models
### windspeed with handheld anemometer
```{r ws ch4 mod}
ch4_ws <- lmer(flux_ch4_g_hr ~ ws_mean + (1|sampling), data = diffusion_env)
summary(ch4_ws)

ch4_ws_b <- lm(flux_ch4_g_hr ~ ws_mean, data = diffusion_env %>% filter(site == "Blue Herron"))
summary(ch4_ws_b)

ch4_ws_c <- lmer(flux_ch4_g_hr ~ ws_mean + (1|sampling), data = diffusion_env %>% filter(site != "Blue Herron"))
summary(ch4_ws_c)
```

### Light 
```{r lux ch4 mod}
ch4_lux <- lmer(flux_ch4_g_hr ~ scale(Lux) + (1|sampling), data = diffusion_env)
summary(ch4_lux)

ch4_lux_b <- lm(flux_ch4_g_hr ~ scale(Lux), data = diffusion_env %>% filter(site == "Blue Herron"))
summary(ch4_lux_b)

ch4_lux_c <- lmer(flux_ch4_g_hr ~ scale(Lux) + (1|sampling), data = diffusion_env %>% filter(site != "Blue Herron"))
summary(ch4_lux_c)
r2_nakagawa(ch4_lux_c)
```

### Temperature
```{r temp ch4 mod}
ch4_temp <- lmer(flux_ch4_g_hr ~ temp + (1|sampling), data = diffusion_env)
summary(ch4_temp)
r2_nakagawa(ch4_temp)

ch4_temp_b <- lm(flux_ch4_g_hr ~ temp, data = diffusion_env %>% filter(site == "Blue Herron"))
summary(ch4_temp_b)

ch4_temp_c <- lmer(flux_ch4_g_hr ~ temp + (1|sampling), data = diffusion_env %>% filter(site != "Blue Herron"))
summary(ch4_temp_c)
```

### Dissolved oxygen
```{r do ch4 mod}
ch4_bottom_do <- lmer(flux_ch4_g_hr ~ mean_bottom_do + (1|sampling), data = diffusion_env)
summary(ch4_bottom_do)

ch4_top_do <- lmer(flux_ch4_g_hr ~ mean_top_do + (1|sampling), data = diffusion_env)
summary(ch4_top_do)
r2_nakagawa(ch4_top_do)
report(ch4_top_do)

diffusion_env %>% group_by(sampling) %>% summarize(slope_percentage_of_range = (0.00084/(max(flux_ch4_g_hr)-min(flux_ch4_g_hr)))*100)
```

### Multiple predictors
```{r}
ch4_all <- lmer(flux_ch4_g_hr ~ scale(mean_top_do) + scale(mean_bottom_do) + scale(temp) + scale(Lux) + scale(ws_mean) + (1|sampling), data = diffusion_env, na.action = "na.fail")
summary(ch4_all)

bh <- diffusion_env %>% filter(site == "Blue Herron")

# how to describe this?
ch4_bh <-lm (flux_ch4_g_hr ~ scale(mean_top_do) + scale(mean_bottom_do) + scale(temp) + scale(Lux) + scale(ws_mean), data = bh, na.action = "na.fail")
dredge(global.model = ch4_all)
dredge(global.model = ch4_bh)
```

## CO2 models
### windspeed with handheld anemometer
```{r ws co2 mod}
co2_ws <- lmer(flux_co2_g_hr ~ ws_mean + (1|sampling), data = diffusion_env)
summary(co2_ws)

co2_ws_b <- lm(flux_co2_g_hr ~ ws_mean, data = diffusion_env %>% filter(site == "Blue Herron"))
summary(co2_ws_b)

co2_ws_c <- lmer(flux_co2_g_hr ~ ws_mean + (1|sampling), data = diffusion_env %>% filter(site != "Blue Herron"))
summary(co2_ws_c)
r2_nakagawa(co2_ws_c)
```

### Light 
```{r lux co2 mod}
co2_lux <- lmer(flux_co2_g_hr ~ scale(Lux) + (1|sampling), data = diffusion_env)
summary(co2_lux)

co2_lux_b <- lm(flux_co2_g_hr ~ scale(Lux), data = diffusion_env %>% filter(site == "Blue Herron"))
summary(co2_lux_b)

co2_lux_c <- lmer(flux_co2_g_hr ~ scale(Lux) + (1|sampling), data = diffusion_env %>% filter(site != "Blue Herron"))
summary(co2_lux_c)
r2_nakagawa(co2_lux_c)
```

### Temperature
```{r temp co2 mod}
co2_temp <- lmer(flux_co2_g_hr ~ temp + (1|sampling), data = diffusion_env)
summary(co2_temp)
r2_nakagawa(co2_temp)
report(co2_temp)

co2_temp_b <- lm(flux_co2_g_hr ~ temp, data = diffusion_env %>% filter(site == "Blue Herron"))
summary(co2_temp_b)

co2_temp_c <- lmer(flux_co2_g_hr ~ temp + (1|sampling), data = diffusion_env %>% filter(site != "Blue Herron"))
summary(co2_temp_c)
r2_nakagawa(co2_temp_c)
report(co2_temp_c)

diffusion_env %>% filter(site != "Blue Herron") %>% group_by(sampling) %>% summarize(slope_percentage_of_range = (0.0337/(max(flux_co2_g_hr)-min(flux_co2_g_hr)))*100)
```

### Dissolved oxygen
```{r do co2 mod}
co2_bottom_do <- lmer(flux_co2_g_hr ~ mean_bottom_do + (1|sampling), data = diffusion_env)
summary(co2_bottom_do)

co2_top_do <- lmer(flux_co2_g_hr ~ mean_top_do + (1|sampling), data = diffusion_env)
summary(co2_top_do)
r2_nakagawa(co2_top_do)
report(co2_top_do)
```

### Multiple predictors
```{r co2 all pred}
co2_all <- lmer(flux_co2_g_hr ~ scale(mean_top_do) + scale(mean_bottom_do) + scale(temp) + scale(Lux) + scale(ws_mean) + (1|sampling), data = diffusion_env, na.action = "na.fail")
summary(co2_all)

dredge(global.model = co2_all)
```

## CH4 plots
```{r ch4 env plots}
ggplotly(ggplot(diffusion_env, aes(temp, flux_ch4_g_hr, color = sampling)) + geom_point()+geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(mean_bottom_do, flux_ch4_g_hr, color = sampling)) + geom_point()+geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(mean_top_do, flux_ch4_g_hr, color = sampling)) + geom_point()+ geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(Lux, flux_ch4_g_hr, color = sampling)) + geom_point()+ geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(ws_mean, flux_ch4_g_hr, color = sampling)) + geom_point()+ geom_smooth(method = "lm"))
```


```{r env mod figs}
pred.ch4.do <- ggpredict(ch4_top_do, terms =c("mean_top_do", "sampling"), type = "random") 

facet.lab <- c("Blue Herron\n13-14 Sept", "Catfish\n06-07 Sept", "Catfish\n18-19 Sept", "Deans\n16-17 Aug","Deans\n30-31 Aug", "Sister\n22-23 Aug")
names(facet.lab) <- c("Blue Herron 13-14 Sept", "Catfish 06-07 Sept", "Catfish 18-19 Sept", "Deans 16-17 Aug","Deans 30-31 Aug", "Sister 22-23 Aug")

#png(filename = "4-Figures/do-ch4.png", width = 8, height = 4, units = "in", res = 300)
ggplot(pred.ch4.do) + 
  geom_ribbon(aes(x = x, ymin = conf.low, ymax = conf.high), alpha = 0.2)+ 
  geom_point(data = diffusion_env %>% rename("group" = "sampling"), aes(mean_top_do, flux_ch4_g_hr, group = group))+ 
  geom_line(aes(x=x, y=predicted))+
  facet_grid(~group, labeller = labeller(group = facet.lab))+
  ylab(expression(Diffusive ~ CH[4]~ flux ~ (g~m^-2~hr^-1)))+
  xlab("Dissolved Oxygen (mg/L)")+
  theme_bw()+ 
  theme(legend.title = element_blank(), axis.text = element_text(size = 14, color = "black"), axis.title = element_text(size = 16, color = "black"), strip.text = element_text(size = 14,  color = "black"))
#dev.off()

pred.co2.temp <- ggpredict(co2_temp, terms =c("temp", "sampling"), type = "random") 

png(filename = "4-Figures/temp-co2.png", width = 8, height = 4, units = "in", res = 300)
ggplot(pred.co2.temp) + 
  geom_ribbon(aes(x = x, ymin = conf.low, ymax = conf.high), alpha = 0.2)+ 
  geom_point(data = diffusion_env %>% rename("group" = "sampling"), aes(temp, flux_co2_g_hr, group = group))+ 
  geom_line(aes(x=x, y=predicted))+
  facet_grid(~group, labeller = labeller(group = facet.lab))+
  ylab(expression(Diffusive ~ CO[2]~ flux ~ (g~m^-2~hr^-1)))+
  scale_x_continuous(breaks = c(20,24,28))+
  xlab("Temperature (°C)")+
  theme_bw()+ 
  theme(legend.title = element_blank(), axis.text = element_text(size = 14, color = "black"), axis.title = element_text(size = 16, color = "black"), strip.text = element_text(size = 14,  color = "black"))
dev.off()

pred.co2.do <- ggpredict(co2_top_do, terms =c("mean_top_do", "sampling"), type = "random") 

png(filename = "4-Figures/do-co2.png", width = 8, height = 4, units = "in", res = 300)
ggplot(pred.co2.do) + 
  geom_ribbon(aes(x = x, ymin = conf.low, ymax = conf.high), alpha = 0.2)+ 
  geom_point(data = diffusion_env %>% rename("group" = "sampling"), aes(mean_top_do, flux_co2_g_hr, group = group))+ 
  geom_line(aes(x=x, y=predicted))+
  facet_grid(~group, labeller = labeller(group = facet.lab))+
  ylab(expression(Diffusive ~ CO[2]~ flux ~ (g~m^-2~hr^-1)))+
  xlab("Dissolved Oxygen (mg/L)")+
  theme_bw()+ 
  theme(legend.title = element_blank(), axis.text = element_text(size = 14, color = "black"), axis.title = element_text(size = 16, color = "black"), strip.text = element_text(size = 14,  color = "black"))
dev.off()

```


## CO2 plots
```{r co2 env plots}
ggplotly(ggplot(diffusion_env, aes(temp, flux_co2_g_hr, color = sampling)) + geom_point()+geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(mean_bottom_do, flux_co2_g_hr, color = sampling)) + geom_point()+geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(mean_top_do, flux_co2_g_hr, color = sampling)) + geom_point()+ geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(Lux, flux_co2_g_hr, color = sampling)) + geom_point()+ geom_smooth(method = "lm"))

ggplotly(ggplot(diffusion_env, aes(ws_mean, flux_co2_g_hr, color = sampling)) + geom_point()+ geom_smooth(method = "lm"))

png(filename = "4-Figures/temp-co2.png", width = 8, height = 5, units = "in", res = 300)
ggplot(diffusion_env, aes(temp, flux_co2_g_hr, color = sampling)) + geom_point()+geom_smooth(method = "lm")+labs(x = "Temperature (°C)", y = "CO2 Diffusion (g/m2/hr)") + theme_bw()
dev.off()
```

```{r raw values diel, echo = F}
ggplot(diffusion_env, aes(trip, flux_ch4_g_hr, group = trip))+geom_point()+ geom_boxplot() + facet_wrap(~sampling)

ggplot(diffusion_env %>% filter(site == "Blue Herron"), aes(start_time, flux_ch4_g_hr, group = trip))+geom_point()

ggplot(diffusion_env %>% filter(site != "Blue Herron"), aes(trip, flux_co2_g_hr, group = trip))+geom_point()+ geom_boxplot() + facet_wrap(~sampling)
```

```{r}
k <- 8.617e-5
stand <- 273.15 + 12


diffusion_env %>% mutate(temp_k = temp + 273.15, botz = (1/(temp_k * k))) %>% ggplot(aes(temp, flux_co2_g_hr)) + geom_point() + geom_smooth(method = "lm") + facet_wrap(~sampling, scales = "free_x")
```


```{r ws models, include = F}
# png(filename = "4-Figures/ws-ch4.png", width = 8, height = 5, units = "in", res = 300)
# ggplot(diffusion_ws, aes(ws_mean, flux_ch4_g_hr, color = sampling, fill = sampling)) +
#   geom_point(shape = 1, size = 2.5, stroke = 1)+
#   geom_smooth(method = "lm")+
#   scale_color_manual(values = c("#DCE319FF","#CD984B","#55C667FF", "#143C1B", "#440154FF", "#2D708EFF"))+
#   scale_fill_manual(values = c("#DCE319FF","#CD984B","#55C667FF", "#143C1B", "#440154FF", "#2D708EFF"))+
#   xlab(expression(Wind ~ Speed ~ (m ~ s^-1)))+
#   ylab(expression(Diffusive ~ Flux ~ (g~CH[4]~m^-2~hr^-1)))+
#   theme_bw()+
#   theme(axis.text = element_text(size = 14, color = "black"), 
#         axis.title = element_text(size = 16, color = "black"), strip.text = element_text(size = 14,  color = "black"), legend.title = element_blank(), legend.text = element_text(size = 14, color = "black"))
#  dev.off()
# 
# png(filename = "4-Figures/ws-co2.png", width = 8, height = 5, units = "in", res = 300)
# ggplot(diffusion_ws, aes(ws_mean, flux_co2_g_hr, color = sampling, fill = sampling)) +
#   geom_point(shape = 1, size = 2.5, stroke = 1)+
#   geom_smooth(method = "lm")+
#   scale_color_manual(values = c("#DCE319FF","#CD984B","#55C667FF", "#143C1B", "#440154FF", "#2D708EFF"))+
#   scale_fill_manual(values = c("#DCE319FF","#CD984B","#55C667FF", "#143C1B", "#440154FF", "#2D708EFF"))+
#   xlab(expression(Wind ~ Speed ~ (m ~ s^-1)))+
#   ylab(expression(Diffusive ~ Flux ~ (g~CO[2]~m^-2~hr^-1)))+
#   theme_bw()+
#   theme(axis.text = element_text(size = 14, color = "black"), 
#         axis.title = element_text(size = 16, color = "black"), strip.text = element_text(size = 14,  color = "black"), legend.title = element_blank(), legend.text = element_text(size = 14, color = "black"))
#  dev.off()
```
```{r mixed model random effect, include = F}
# df1 <- data.frame(x = c(5, 4.5, 4, 3, 2.5, 2, 2, 1.5, 1), y = c(0,0.5,1, 2, 2.5, 3, 4, 4.5, 5), z = c(rep("a", 3), rep("b", 3), rep("c", 3)))
# 
# df2 <- data.frame(x = c(5, 4.5, 4, 3, 2.5, 2, 2, 1.5, 1), y = c( 2, 2.5, 3, 4, 4.5, 5, 0,0.5,1), z = c(rep("a", 3), rep("b", 3), rep("c", 3)))
# 
# ggplot(df2, aes(x, y, color = z))+geom_point()
# 
# summary(lmer(y ~ x + (1|z), data = df1))
# summary(lmer(y~x + (1|z), data = df2))
```
